check if the resource is new then the delete is noop

+++ src/main/java/org/springframework/data/dozer/repository/support/SimpleDozerRepository.java
@@ -465,6 +465,10 @@ public class SimpleDozerRepository<T, ID> implements DozerRepositoryImplementati
 
 	@Override
 	public void delete(T resource) {
+		if (entityInformation.isNew(resource)) {
+			return;
+		}
+		
 		Object entityId = toAdaptedId((ID) entityInformation.getId(resource));
 
 		getAdaptedRepository().deleteById(entityId);
